Enable typescript support and add path for migrating application: * Organised the repo into a more npm friendly dir structure * Added linters * Migrated index of application to typescript
